Finished Testing on Ryan Wang Zifeng's MEDCLIP
it seems that Ryan Wang Zifeng Et Al did not train the their model on ROCO, they utilized: CheXpert, MIMIC-CXR, COVID and RSNA Pneumonia datasets. While the initial format of our datasets seem similar, this model might not have been the best choice to try first. For my next experiment, I will be trying the Kaushalya MEDCLIP which was trained on the ROCO 1 dataset. But I will be using ROCO 2 which should be similar enough.